+++ src/components/common/CommSpinner.js
@@ -1,0 +1,7 @@
+import PulseLoader from "react-spinners/PulseLoader";
+
+const CommSpinner = () => {
+  return <PulseLoader size={10} color="#000080" />;
+};
+
+export default CommSpinner;